+++ src/pages/profile/order-lines/order-lines.html
@@ -17,7 +17,7 @@
     <ion-card *ngFor="let ol of noDuplicateOrderLine">
         <ion-row style="padding: 5px">
             <img src="{{getThumbnailURL(ol.boughtColor, ol.product)}}"/>
-            <div style="color: black; width: 100%; font-weight: bold; text-align: left">
+            <div style="color: black; width: 100%; font-weight: bold;">
                 {{ol.product.name}}
             </div>
         </ion-row>
