+++ src/services/order.service.ts
@@ -1,13 +1,12 @@
 import {Injectable} from '@angular/core';
 import {BehaviorSubject, Subject} from 'rxjs';
-import {IPageInfo} from '../interfaces/ipageInfo.interface';
 import {HttpService} from './http.service';
 import {ToastController} from 'ionic-angular';
 
 @Injectable()
 export class OrderService {
   // orderArray: BehaviorSubject<any> = new BehaviorSubject<any>([]);
-  orderArray: any = [];
+  orderArray: BehaviorSubject<any> = new BehaviorSubject<any>([]);
   orderData: any;
   constructor(private httpService: HttpService, private toastCtrl: ToastController) {
   }
@@ -15,7 +14,7 @@ export class OrderService {
   getAllOrders() {
     this.httpService.get('orders').subscribe(
       (info) => {
-        this.orderArray = info.orders;
+        this.orderArray.next(info.orders);
       },
       (err) => {
         console.error('error');
