+++ src/shared/lib/makePersianNumber.ts
@@ -1,0 +1,5 @@
+export function makePersianNumber(a: string, isPrice) {
+  if (isNaN((+a)))
+    return a;
+  return (+a).toLocaleString('fa', {useGrouping: isPrice});
+}
